guestbook #1

+++ mysite03/src/main/java/com/douzone/mysite/vo/GuestbookVo.java
@@ -1,15 +1,15 @@
 package com.douzone.mysite.vo;
 
 public class GuestbookVo {
-	private int no;
+	private Long no;
 	private String name;
 	private String password;
 	private String message;
 	private String RegDate;
-	public int getNo() {
+	public Long getNo() {
 		return no;
 	}
-	public void setNo(int no) {
+	public void setNo(Long no) {
 		this.no = no;
 	}
 	public String getName() {
